+++ pkg/client/api/v0040/jobInfo.go
@@ -8,6 +8,9 @@ import (
 	"errors"
 	"net/http"
 
+	utilerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/utils/ptr"
+
 	api "github.com/SlinkyProject/slurm-client/api/v0040"
 	"github.com/SlinkyProject/slurm-client/pkg/types"
 	"github.com/SlinkyProject/slurm-client/pkg/utils"
@@ -31,7 +34,15 @@ func (c *SlurmClient) CreateJobInfo(ctx context.Context, req any) (*int32, error
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	}
 	return res.JSON200.JobId, nil
 }
@@ -43,7 +54,15 @@ func (c *SlurmClient) DeleteJobInfo(ctx context.Context, jobId string) error {
 	if err != nil {
 		return err
 	} else if res.StatusCode() != 200 {
-		return errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return utilerrors.NewAggregate(errs)
 	}
 	return nil
 }
@@ -55,7 +74,15 @@ func (c *SlurmClient) GetJobInfo(ctx context.Context, jobId string) (*types.V004
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	} else if len(res.JSON200.Jobs) == 0 {
 		return nil, errors.New(http.StatusText(http.StatusNotFound))
 	}
@@ -71,7 +98,15 @@ func (c *SlurmClient) ListJobInfo(ctx context.Context) (*types.V0040JobInfoList,
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	}
 	list := &types.V0040JobInfoList{
 		Items: make([]types.V0040JobInfo, len(res.JSON200.Jobs)),