+++ pkg/client/api/v0040/node.go
@@ -8,6 +8,9 @@ import (
 	"errors"
 	"net/http"
 
+	utilerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/utils/ptr"
+
 	api "github.com/SlinkyProject/slurm-client/api/v0040"
 	"github.com/SlinkyProject/slurm-client/pkg/types"
 	"github.com/SlinkyProject/slurm-client/pkg/utils"
@@ -42,7 +45,15 @@ func (c *SlurmClient) UpdateNode(ctx context.Context, nodeName string, req any) 
 	if err != nil {
 		return err
 	} else if res.StatusCode() != 200 {
-		return errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return utilerrors.NewAggregate(errs)
 	}
 	return nil
 }
@@ -54,7 +65,15 @@ func (c *SlurmClient) GetNode(ctx context.Context, nodeName string) (*types.V004
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	} else if len(res.JSON200.Nodes) == 0 {
 		return nil, errors.New(http.StatusText(http.StatusNotFound))
 	}
@@ -70,7 +89,15 @@ func (c *SlurmClient) ListNodes(ctx context.Context) (*types.V0040NodeList, erro
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	}
 	nodeList := &types.V0040NodeList{
 		Items: make([]types.V0040Node, len(res.JSON200.Nodes)),