+++ pkg/client/api/v0040/partitionInfo.go
@@ -8,6 +8,9 @@ import (
 	"errors"
 	"net/http"
 
+	utilerrors "k8s.io/apimachinery/pkg/util/errors"
+	"k8s.io/utils/ptr"
+
 	api "github.com/SlinkyProject/slurm-client/api/v0040"
 	"github.com/SlinkyProject/slurm-client/pkg/types"
 	"github.com/SlinkyProject/slurm-client/pkg/utils"
@@ -25,7 +28,15 @@ func (c *SlurmClient) GetPartitionInfo(ctx context.Context, name string) (*types
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	} else if len(res.JSON200.Partitions) == 0 {
 		return nil, errors.New(http.StatusText(http.StatusNotFound))
 	}
@@ -41,7 +52,15 @@ func (c *SlurmClient) ListPartitionInfo(ctx context.Context) (*types.V0040Partit
 	if err != nil {
 		return nil, err
 	} else if res.StatusCode() != 200 {
-		return nil, errors.New(http.StatusText(res.StatusCode()))
+		errs := []error{errors.New(http.StatusText(res.StatusCode()))}
+		if res.JSONDefault != nil {
+			for _, e := range ptr.Deref(res.JSONDefault.Errors, []api.V0040OpenapiError{}) {
+				if e.Error != nil {
+					errs = append(errs, errors.New(*e.Error))
+				}
+			}
+		}
+		return nil, utilerrors.NewAggregate(errs)
 	}
 	list := &types.V0040PartitionInfoList{
 		Items: make([]types.V0040PartitionInfo, len(res.JSON200.Partitions)),