+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Added `Create()` and `Delete()` support for v40 JobInfo.
 - Added `Create()` and `Delete()` support for v41 JobInfo, including
   "placeholder" job.
+- Added aggregated errors created from Slurm API errors.
 
 ### Fixed